+++ common/src/main/java/club/iananderson/seasonhud/impl/seasons/Calendar.java
@@ -31,4 +31,19 @@ public class Calendar {
     return findCalendar(mc.player, calendar) || !Config.getNeedCalendar();
   }
 
+  public static boolean calendarFoundDetailed() {
+    Minecraft mc = Minecraft.getInstance();
+    Item calendar = Services.SEASON.calendar();
+
+    if (Common.platformName().equals("Fabric") && !Common.extrasLoaded()) {
+      return false;
+    }
+
+    if (mc.level == null || mc.player == null || calendar == null) {
+      return false;
+    }
+
+    return findCalendar(mc.player, calendar) && Config.getCalanderDetailMode();
+  }
+
 }